+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+- [x] __[VPN Commander]__ connect vpn services from Alfred.
+
 # v3
 __[Yandex Translate]__ Yandex Dictionary support https://yandex.ru/dev/dictionary/doc/dg/reference/lookup.html.